La screen de mirar rutina de la concha de la re puta lora

+++ app/src/main/java/com/grupo14/gym_androidapp/api/models/RoutineApiModel.kt
@@ -6,7 +6,7 @@ import java.util.Date
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 data class RoutineApiModel (
-    @JsonProperty("id") val id: Long?,
+    @JsonProperty("id") val id: Int?,
     @JsonProperty("name") val name: String?,
     @JsonProperty("detail") val detail: String?,
     @JsonProperty("date") val date: Date?,
