+++ app/src/main/java/com/grupo14/gym_androidapp/api/models/SmallUserApiModel.kt
@@ -6,7 +6,7 @@ import java.util.*
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 data class SmallUserApiModel(
-    @JsonProperty("id") val id: Long?,
+    @JsonProperty("id") val id: Int?,
     @JsonProperty("username") val username: String?,
     @JsonProperty("gender") val gender: Gender?,
     @JsonProperty("avatarUrl") val avatarUrl: String?,
